Example notebook and class modification
Created an examples folder with a notebook 'usage.ipynb' which shows some of the base features of the package.

Modified the Atomic_Embeddings class to handle linear representations and to reduce the redundant lines to load in the representations included within the package.

Modified the functions which referenced the element_list() method to reference the element_list class property.

Move the data folder to be within the package directory.